Moved pkg/core/wharfyml to pkg/wharfyml
