+++ pkg/wharfyml/magicstrings.go
