+++ pkg/worker/worker.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/iver-wharf/wharf-cmd/pkg/core/wharfyml"
+	"github.com/iver-wharf/wharf-cmd/pkg/wharfyml"
 	"github.com/iver-wharf/wharf-core/pkg/logger"
 )
